+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/shared_component/InfoSection.kt
@@ -1,4 +1,4 @@
-package com.sanaa.presentation.component
+package com.sanaa.presentation.shared_component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
