+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/shared_component/RequestToLoginBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.sanaa.presentation.component
+package com.sanaa.presentation.shared_component
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.isSystemInDarkTheme
@@ -48,7 +48,7 @@ fun RequestToLoginBottomSheet(
         onDismiss = onDismiss,
         content = {
             Column(
-                modifier = modifier.fillMaxWidth(),
+                modifier = modifier.fillMaxWidth().padding(bottom = 24.dp),
                 horizontalAlignment = Alignment.CenterHorizontally,
             ) {
                 NovixTopBar(
