+++ feature/media_details/presentation/src/main/java/com/sanaa/presentation/shared_component/cards/MediaPosterCard.kt
@@ -1,4 +1,4 @@
-package com.sanaa.presentation.component.cards
+package com.sanaa.presentation.shared_component.cards
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
